refactor: re-organize package and fix imports
