+++ src/main/scala/io/github/srs/model/entity/dynamic_entity/Actuator.scala
@@ -1,4 +1,4 @@
-package io.github.srs.model.entity.dynamic
+package io.github.srs.model.entity.dynamic_entity
 
 enum Actuator:
   case WheelMotor(left: Wheel, right: Wheel)
