+++ src/test/scala/io/github/srs/model/entity/OrientationTest.scala
@@ -1,4 +1,4 @@
-package io.github.srs.model
+package io.github.srs.model.entity
 
 import io.github.srs.model.entity.Orientation
 import org.scalatest.flatspec.AnyFlatSpec
