+++ src/test/scala/io/github/srs/model/entity/Point2DTest.scala
@@ -1,4 +1,4 @@
-package io.github.srs.model
+package io.github.srs.model.entity
 
 import io.github.srs.model.entity.*
 import org.scalatest.flatspec.AnyFlatSpec
