+++ src/test/scala/io/github/srs/model/entity/ShapeTypeTest.scala
@@ -1,9 +1,9 @@
-package io.github.srs.model
+package io.github.srs.model.entity
 
 import io.github.srs.model.entity.ShapeType
+import io.github.srs.model.entity.ShapeType.*
 import org.scalatest.flatspec.AnyFlatSpec
 import org.scalatest.matchers.should
-import io.github.srs.model.entity.ShapeType.*
 
 class ShapeTypeTest extends AnyFlatSpec with should.Matchers:
